requirements.txt
langchain==0.0.137
python-dotenv==1.0.0
streamlit==1.18.1
openai==0.27.8
tiktoken==0.4.0
altair==4